# internal/server/api_kanban.go
// Package server - kanban board API handlers
package server

import (
	"encoding/json"
	"fmt"
	"net/http"
	"slices"
	"strconv"
	"strings"
	"time"

	"knov/internal/configmanager"
	"knov/internal/filter"
	"knov/internal/kanban"
	"knov/internal/logging"
	"knov/internal/pathutils"
	"knov/internal/server/notify"
	"knov/internal/server/render"
	"knov/internal/translation"

	"github.com/go-chi/chi/v5"
)

// resolveBoard looks up a configured kanban board by its URL slug, writing a 404 if unknown.
func resolveBoard(w http.ResponseWriter, r *http.Request) (configmanager.KanbanBoard, bool) {
	slug := chi.URLParam(r, "board")
	if slug == "" {
		http.Error(w, translation.SprintfForRequest(configmanager.GetLanguage(), "missing board"), http.StatusBadRequest)
		return configmanager.KanbanBoard{}, false
	}
	board, ok := configmanager.GetKanbanBoardBySlug(slug)
	if !ok {
		http.Error(w, translation.SprintfForRequest(configmanager.GetLanguage(), "unknown board"), http.StatusNotFound)
		return configmanager.KanbanBoard{}, false
	}
	return board, true
}

// @Summary Get kanban board for a folder
// @Description Returns all kanban cards grouped by status column for the given board
// @Tags kanban
// @Param board path string true "Board slug"
// @Param ancestor query string false "Filter by ancestor (epic)"
// @Param tag query string false "Filter by tag"
// @Param q query string false "Search query"
// @Produce json,html
// @Router /api/kanban/{board} [get]
func handleAPIGetKanbanBoard(w http.ResponseWriter, r *http.Request) {
	board, ok := resolveBoard(w, r)
	if !ok {
		return
	}

	cfg := &filter.Config{Logic: "and"}

	if ancestor := r.URL.Query().Get("ancestor"); ancestor != "" {
		cfg.Criteria = append(cfg.Criteria, filter.Criteria{Metadata: "ancestor-of", Operator: "equals", Value: ancestor, Action: "include"})
	}
	if tag := r.URL.Query().Get("tag"); tag != "" {
		cfg.Criteria = append(cfg.Criteria, filter.Criteria{Metadata: "tags", Operator: "equals", Value: tag, Action: "include"})
	}

	cols, _ := kanban.BuildBoard(board.FolderPath, cfg, strings.ToLower(r.URL.Query().Get("q")), kanban.SortBy(r.URL.Query().Get("sort")))
	writeResponse(w, r, cols, render.RenderKanbanBoard(cols))
}

// @Summary Get archived kanban cards for a board
// @Description Returns all cards with the archive status for the board's folder (and subfolders), newest first
// @Tags kanban
// @Param board path string true "Board slug"
// @Produce json,html
// @Router /api/kanban/{board}/archive [get]
func handleAPIGetKanbanArchive(w http.ResponseWriter, r *http.Request) {
	board, ok := resolveBoard(w, r)
	if !ok {
		return
	}

	cards, err := kanban.Archived(board.FolderPath)
	if err != nil {
		logging.LogError(logging.KeyApp, "failed to get archived cards for %s: %v", board.FolderPath, err)
		http.Error(w, translation.SprintfForRequest(configmanager.GetLanguage(), "failed to get archived cards"), http.StatusInternalServerError)
		return
	}

	writeResponse(w, r, cards, render.RenderKanbanArchive(cards))
}

// @Summary Apply advanced filter to kanban board
// @Description Filters the kanban board using the full filter form, scoped to the board's folder
// @Tags kanban
// @Accept application/x-www-form-urlencoded
// @Produce json,html
// @Param board path string true "Board slug"
// @Success 200 {string} string "kanban board html"
// @Router /api/kanban/{board}/filter [post]
func handleAPIPostKanbanFilter(w http.ResponseWriter, r *http.Request) {
	board, ok := resolveBoard(w, r)
	if !ok {
		return
	}

	if err := r.ParseForm(); err != nil {
		http.Error(w, translation.SprintfForRequest(configmanager.GetLanguage(), "failed to parse form"), http.StatusBadRequest)
		return
	}

	cfg := filter.ParseFilterConfigFromForm(r, -1)
	cfg.Logic = "and"

	cols, _ := kanban.BuildBoard(board.FolderPath, cfg, "", kanban.SortBy(r.FormValue("sort")))
	writeResponse(w, r, cols, render.RenderKanbanBoard(cols))
}

// @Summary Move a kanban card to a new status column
// @Description Updates the kanban status tag on a file, replacing any existing kanban tag
// @Tags kanban
// @Accept application/x-www-form-urlencoded
// @Produce json,html
// @Param filepath formData string true "File path"
// @Param status formData string true "New kanban status"
// @Param board formData string false "Board slug (scopes the event log entry; omit to guess from the file's folder)"
// @Success 200 {string} string "card updated"
// @Failure 400 {string} string "missing parameter"
// @Failure 404 {string} string "file not found"
// @Failure 500 {string} string "failed to update"
// @Router /api/kanban/card/move [post]
func handleAPIKanbanMoveCard(w http.ResponseWriter, r *http.Request) {
	if err := r.ParseForm(); err != nil {
		http.Error(w, translation.SprintfForRequest(configmanager.GetLanguage(), "failed to parse form"), http.StatusBadRequest)
		return
	}

	filePath := r.FormValue("filepath")
	newStatus := r.FormValue("status")

	if filePath == "" || newStatus == "" {
		http.Error(w, translation.SprintfForRequest(configmanager.GetLanguage(), "missing filepath or status"), http.StatusBadRequest)
		return
	}
	if !slices.Contains(configmanager.GetKanbanStatuses(), newStatus) {
		http.Error(w, translation.SprintfForRequest(configmanager.GetLanguage(), "invalid kanban status"), http.StatusBadRequest)
		return
	}

	var boardFolder string
	if board, ok := configmanager.GetKanbanBoardBySlug(r.FormValue("board")); ok {
		boardFolder = board.FolderPath
	}

	oldStatus, err := kanban.MoveCard(boardFolder, filePath, newStatus)
	if err != nil {
		logging.LogError(logging.KeyApp, "failed to move kanban card %s to %s: %v", filePath, newStatus, err)
		notify.SetHeader(w, notify.LevelError, translation.SprintfForRequest(configmanager.GetLanguage(), "failed to update card"))
		http.Error(w, translation.SprintfForRequest(configmanager.GetLanguage(), "failed to update card"), http.StatusInternalServerError)
		return
	}

	var msg string
	if oldStatus == "" {
		msg = translation.SprintfForRequest(configmanager.GetLanguage(), "status added: %s", newStatus)
	} else {
		msg = translation.SprintfForRequest(configmanager.GetLanguage(), "status changed: %s → %s", oldStatus, newStatus)
	}
	notify.SetHeader(w, notify.LevelSuccess, msg)
	writeResponse(w, r, map[string]string{"filepath": filePath, "status": newStatus}, "")
}

// @Summary Save card order for a kanban column
// @Description Persists the drag-and-drop card order for all columns in a board
// @Tags kanban
// @Accept application/x-www-form-urlencoded
// @Param board path string true "Board slug"
// @Param status formData string true "Column status"
// @Param order formData string true "Comma-separated list of filepaths in display order"
// @Success 200 {string} string "order saved"
// @Router /api/kanban/{board}/order [post]
func handleAPIKanbanSaveOrder(w http.ResponseWriter, r *http.Request) {
	board, ok := resolveBoard(w, r)
	if !ok {
		return
	}
	if err := r.ParseForm(); err != nil {
		http.Error(w, translation.SprintfForRequest(configmanager.GetLanguage(), "failed to parse form"), http.StatusBadRequest)
		return
	}

	status := r.FormValue("status")
	if status == "" {
		http.Error(w, translation.SprintfForRequest(configmanager.GetLanguage(), "missing status"), http.StatusBadRequest)
		return
	}

	var paths []string
	for _, p := range strings.Split(r.FormValue("order"), ",") {
		if p = strings.TrimSpace(p); p != "" {
			paths = append(paths, p)
		}
	}

	stored, err := kanban.GetOrder(board.FolderPath)
	if err != nil {
		logging.LogError(logging.KeyApp, "kanban: load order failed for %s: %v", board.FolderPath, err)
		stored = kanban.Order{}
	}
	stored[status] = paths

	if err := kanban.SaveOrder(board.FolderPath, stored); err != nil {
		logging.LogError(logging.KeyApp, "kanban: save order failed for %s: %v", board.FolderPath, err)
		http.Error(w, translation.SprintfForRequest(configmanager.GetLanguage(), "failed to save order"), http.StatusInternalServerError)
		return
	}

	w.WriteHeader(http.StatusOK)
}

// @Summary Get a short text excerpt from a file
// @Description Returns the first N runes of meaningful body text, stripping front matter and markdown syntax
// @Tags kanban
// @Param filepath query string true "File path (relative)"
// @Param chars query int false "Max runes to return (default 30)"
// @Produce html
// @Success 200 {string} string "excerpt text"
// @Router /api/kanban/excerpt [get]
func handleAPIGetKanbanExcerpt(w http.ResponseWriter, r *http.Request) {
	filePath := r.URL.Query().Get("filepath")
	if filePath == "" {
		writeResponse(w, r, map[string]string{"excerpt": ""}, "")
		return
	}

	chars := 30
	if c := r.URL.Query().Get("chars"); c != "" {
		if n, err := strconv.Atoi(c); err == nil && n > 0 {
			chars = n
		}
	}

	excerpt := kanban.Excerpt(pathutils.ToDocsPath(filePath), chars)

	var html string
	if excerpt != "" {
		html = fmt.Sprintf(`<div class="kanban-card-excerpt">%s</div>`, excerpt)
	}
	writeResponse(w, r, map[string]string{"excerpt": excerpt}, html)
}

// @Summary Get all non-kanban tags used in a board's kanban cards
// @Tags kanban
// @Param board path string true "Board slug"
// @Produce html
// @Router /api/kanban/{board}/tags [get]
func handleAPIGetKanbanTags(w http.ResponseWriter, r *http.Request) {
	board, ok := configmanager.GetKanbanBoardBySlug(chi.URLParam(r, "board"))
	if !ok {
		writeResponse(w, r, []string{}, "")
		return
	}

	tags, err := kanban.TagsForFolder(board.FolderPath)
	if err != nil {
		writeResponse(w, r, []string{}, "")
		return
	}

	var html strings.Builder
	for _, t := range tags {
		fmt.Fprintf(&html, `<option value="%s">%s</option>`, t, t)
	}
	writeResponse(w, r, tags, html.String())
}

// @Summary Get kanban event log
// @Description Returns kanban card move events, newest first. All parameters are optional.
// @Tags kanban
// @Param board path string true "Board slug"
// @Param file query string false "Filter by file path"
// @Param from query string false "Start of time range (RFC3339 or YYYY-MM-DD)"
// @Param to query string false "End of time range (RFC3339 or YYYY-MM-DD)"
// @Param limit query int false "Max number of events (default 200, 0 = unlimited)"
// @Produce json,html
// @Router /api/kanban/{board}/events [get]
func handleAPIGetKanbanEvents(w http.ResponseWriter, r *http.Request) {
	board, ok := resolveBoard(w, r)
	if !ok {
		return
	}

	filePath := r.URL.Query().Get("file")
	fromRaw := r.URL.Query().Get("from")
	toRaw := r.URL.Query().Get("to")

	var from, to *time.Time
	if fromRaw != "" {
		if t, err := parseEventBoundary(fromRaw, false); err == nil {
			from = &t
		}
	}
	if toRaw != "" {
		if t, err := parseEventBoundary(toRaw, true); err == nil {
			to = &t
		}
	}

	limit := 200
	if l := r.URL.Query().Get("limit"); l != "" {
		if n, err := strconv.Atoi(l); err == nil && n >= 0 {
			limit = n
		}
	}

	events, err := kanban.GetEvents(board.FolderPath, filePath, from, to, limit)
	if err != nil {
		logging.LogError(logging.KeyApp, "failed to get kanban events for %s: %v", board.FolderPath, err)
		http.Error(w, translation.SprintfForRequest(configmanager.GetLanguage(), "failed to get events"), http.StatusInternalServerError)
		return
	}

	filePaths, err := kanban.FilesForFolder(board.FolderPath)
	if err != nil {
		logging.LogError(logging.KeyApp, "failed to get kanban files for %s: %v", board.FolderPath, err)
	}

	writeResponse(w, r, events, render.RenderKanbanEvents(events, filePaths, board.Slug, filePath, fromRaw, toRaw))
}

// parseEventBoundary parses a time-range boundary as RFC3339, falling back to a bare
// YYYY-MM-DD date (as produced by a native <input type="date">) expanded to the start
// or end of that day.
func parseEventBoundary(s string, endOfDay bool) (time.Time, error) {
	if t, err := time.Parse(time.RFC3339, s); err == nil {
		return t, nil
	}
	d, err := time.Parse("2006-01-02", s)
	if err != nil {
		return time.Time{}, err
	}
	if endOfDay {
		d = d.Add(23*time.Hour + 59*time.Minute + 59*time.Second)
	}
	return d, nil
}

// @Summary Get kanban card file paths
// @Description Returns the file paths of all cards currently on the kanban board for a board, sorted.
// @Tags kanban
// @Param board path string true "Board slug"
// @Produce json
// @Router /api/kanban/{board}/files [get]
func handleAPIGetKanbanFiles(w http.ResponseWriter, r *http.Request) {
	board, ok := resolveBoard(w, r)
	if !ok {
		return
	}

	paths, err := kanban.FilesForFolder(board.FolderPath)
	if err != nil {
		logging.LogError(logging.KeyApp, "failed to get kanban files for %s: %v", board.FolderPath, err)
		http.Error(w, translation.SprintfForRequest(configmanager.GetLanguage(), "failed to get files"), http.StatusInternalServerError)
		return
	}

	w.Header().Set("Content-Type", "application/json")
	json.NewEncoder(w).Encode(paths)
}