# internal/server/swagger/swagger.yaml
basePath: /
...
host: localhost:1324
info:
  contact: {}
  description: KNOV API \n http://localhost:1324
  title: Knov API
  version: "1.0"
paths:
...
  /api/kanban/{board}:
    get:
      description: Returns all kanban cards grouped by status column for the given
        board
      parameters:
      - description: Board slug
        in: path
        name: board
        required: true
        type: string
      - description: Filter by ancestor (epic)
        in: query
        name: ancestor
        type: string
      - description: Filter by tag
        in: query
        name: tag
        type: string
      - description: Search query
        in: query
        name: q
        type: string
      produces:
      - application/json
      - text/html
      responses: {}
      summary: Get kanban board for a folder
      tags:
      - kanban
  /api/kanban/{board}/archive:
    get:
      description: Returns all cards with the archive status for the board's folder
        (and subfolders), newest first
      parameters:
      - description: Board slug
        in: path
        name: board
        required: true
        type: string
      produces:
      - application/json
      - text/html
      responses: {}
      summary: Get archived kanban cards for a board
      tags:
      - kanban
  /api/kanban/{board}/events:
    get:
      description: Returns kanban card move events, newest first. All parameters are
        optional.
      parameters:
      - description: Board slug
        in: path
        name: board
        required: true
        type: string
      - description: Filter by file path
        in: query
        name: file
        type: string
      - description: Start of time range (RFC3339 or YYYY-MM-DD)
        in: query
        name: from
        type: string
      - description: End of time range (RFC3339 or YYYY-MM-DD)
        in: query
        name: to
        type: string
      - description: Max number of events (default 200, 0 = unlimited)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      - text/html
      responses: {}
      summary: Get kanban event log
      tags:
      - kanban
  /api/kanban/{board}/files:
    get:
      description: Returns the file paths of all cards currently on the kanban board
        for a board, sorted.
      parameters:
      - description: Board slug
        in: path
        name: board
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: Get kanban card file paths
      tags:
      - kanban
  /api/kanban/{board}/filter:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: Filters the kanban board using the full filter form, scoped to
        the board's folder
      parameters:
      - description: Board slug
        in: path
        name: board
        required: true
        type: string
      produces:
      - application/json
      - text/html
      responses:
        "200":
          description: kanban board html
          schema:
            type: string
      summary: Apply advanced filter to kanban board
      tags:
      - kanban
  /api/kanban/{board}/order:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: Persists the drag-and-drop card order for all columns in a board
      parameters:
      - description: Board slug
        in: path
        name: board
        required: true
        type: string
      - description: Column status
        in: formData
        name: status
        required: true
        type: string
      - description: Comma-separated list of filepaths in display order
        in: formData
        name: order
        required: true
        type: string
      responses:
        "200":
          description: order saved
          schema:
            type: string
      summary: Save card order for a kanban column
      tags:
      - kanban
  /api/kanban/{board}/tags:
    get:
      parameters:
      - description: Board slug
        in: path
        name: board
        required: true
        type: string
      produces:
      - text/html
      responses: {}
      summary: Get all non-kanban tags used in a board's kanban cards
      tags:
      - kanban
  /api/kanban/card/move:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: Updates the kanban status tag on a file, replacing any existing
        kanban tag
      parameters:
      - description: File path
        in: formData
        name: filepath
        required: true
        type: string
      - description: New kanban status
        in: formData
        name: status
        required: true
        type: string
      - description: Board slug (scopes the event log entry; omit to guess from the
          file's folder)
        in: formData
        name: board
        type: string
      produces:
      - application/json
      - text/html
      responses:
        "200":
          description: card updated
          schema:
            type: string
        "400":
          description: missing parameter
          schema:
            type: string
        "404":
          description: file not found
          schema:
            type: string
        "500":
          description: failed to update
          schema:
            type: string
      summary: Move a kanban card to a new status column
      tags:
      - kanban
  /api/kanban/excerpt:
    get:
      description: Returns the first N runes of meaningful body text, stripping front
        matter and markdown syntax
      parameters:
      - description: File path (relative)
        in: query
        name: filepath
        required: true
        type: string
      - description: Max runes to return (default 30)
        in: query
        name: chars
        type: integer
      produces:
      - text/html
      responses:
        "200":
          description: excerpt text
          schema:
            type: string
      summary: Get a short text excerpt from a file
      tags:
      - kanban
...
swagger: "2.0"